add and delete in dynamic route also

+++ main.js
@@ -73,6 +73,13 @@ const item3 = new Item({
 });
 const defaultItems = [item1, item2, item3];
 
+const listSchema = {
+  name: String,
+  item: [itemSchema],
+};
+
+const List = mongoose.model("list", listSchema);
+
 main.get("/", function (req, res) {
   Item.find({})
     .then((foundItems) => {
@@ -93,26 +100,77 @@ main.get("/", function (req, res) {
       console.error(err);
     });
 });
+main.get("/:dynamicRoute", (req, res) => {
+  const dynamicRoutes = req.params.dynamicRoute;
+  console.log("Dynamic route activated ", dynamicRoutes);
+
+  List.findOne({ name: dynamicRoutes })
+    .then((found) => {
+      if (!found) {
+        const list = new List({
+          name: dynamicRoutes,
+          item: defaultItems,
+        });
+        list.save();
+      } else {
+        res.render("index", {
+          Title: found.name,
+          newListItem: found.item,
+        });
+      }
+    })
+    .catch((error) => {
+      console.log("Couldn't find Route ", error);
+    });
+});
 
 main.post("/", function (req, res) {
   const itemName = req.body.newList;
+  const listName = req.body.submitList;
 
   const item = new Item({
     name: itemName,
   });
-  item.save();
-  res.redirect("/");
+  if (listName === "Today") {
+    item.save();
+    res.redirect("/");
+  } else {
+    List.findOne({ name: listName })
+      .then((found) => {
+        found.item.push(item);
+        found.save();
+        res.redirect("/" + listName);
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  }
 });
 main.post("/delete", function (req, res) {
   const checkBoxId = req.body.checkboxName;
-  Item.findByIdAndDelete(checkBoxId)
-    .then((success) => {
-      console.log("this is a great success!", success);
-      res.redirect("/");
-    })
-    .catch((error) => {
-      console.log("Sorry for the error!");
-    });
+  const hiddenList = req.body.hiddenList;
+
+  if (hiddenList === "Today") {
+    Item.findByIdAndDelete(checkBoxId)
+      .then((success) => {
+        console.log("The item has been deleted ", success);
+        res.redirect("/");
+      })
+      .catch((error) => {
+        console.log("Sorry couldn't delete the item !");
+      });
+  } else {
+    List.findOneAndUpdate(
+      { name: hiddenList },
+      { $pull: { item: { _id: checkBoxId } } }
+    )
+      .then((found) => {
+        res.redirect("/" + hiddenList);
+      })
+      .catch((err) => {
+        console.log("Error " + err);
+      });
+  }
 });
 
 main.listen(4000, () => {
